+++ tater/ui/callbacks/__init__.py
@@ -1,0 +1,54 @@
+"""Dash callback registration for TaterApp.
+
+Re-exports all public setup functions so that ``tater_app.py`` can continue
+to import from ``tater.ui.callbacks`` without changes.
+"""
+from tater.ui.callbacks.core import (
+    setup_callbacks,
+    setup_value_capture_callbacks,
+    _do_capture,
+    _do_load,
+    _decode_field_path,
+)
+from tater.ui.callbacks.span import setup_span_callbacks
+from tater.ui.callbacks.repeater import (
+    setup_repeater_callbacks,
+    setup_nested_repeater_callbacks,
+)
+from tater.ui.callbacks.hl import (
+    setup_hl_callbacks,
+    setup_hl_tags_callbacks,
+)
+from tater.ui.callbacks.helpers import (
+    _default_meta,
+    _get_ann,
+    _get_meta,
+    _has_value,
+    _collect_value_capture_widgets,
+    _collect_all_control_templates,
+    update_status_for_doc,
+    _build_menu_items,
+    _perform_navigation,
+)
+
+__all__ = [
+    "setup_callbacks",
+    "setup_value_capture_callbacks",
+    "setup_span_callbacks",
+    "setup_repeater_callbacks",
+    "setup_nested_repeater_callbacks",
+    "setup_hl_callbacks",
+    "setup_hl_tags_callbacks",
+    "_collect_value_capture_widgets",
+    "_collect_all_control_templates",
+    "_default_meta",
+    "_get_ann",
+    "_get_meta",
+    "_has_value",
+    "update_status_for_doc",
+    "_build_menu_items",
+    "_perform_navigation",
+    "_do_capture",
+    "_do_load",
+    "_decode_field_path",
+]
